Working on BPO Indent creation feature which includes below changes: - Update nodeUser model - update businessNode model - Update registerNode fetch controller - Modify auth controller - Create new Indent Controller - Add two new model Indent and Indent Items and define there associations - update many model structure - working on indent creation controller - Setup new endpoint for indent

+++ src/models/main/rfqItem.model.js
@@ -15,7 +15,7 @@ function RFQItem(sequelize) {
             allowNull: true,
         },
         qty: {
-            type: DataTypes.INTEGER,
+            type: DataTypes.DECIMAL(10, 3),
             allowNull: false,
         },
         uom: {
